Move flag_summary and echo_flag_table to flag.py and fix all_nan_columns definition
