+++ quantaq_cli/console/commands/expunge.py
@@ -3,12 +3,13 @@ from pathlib import Path
 import click
 import numpy as np
 import pandas as pd
-import rich
-from rich.table import Table
 
 from quantaq_cli.exceptions import InvalidFileExtension, InvalidDeviceModel
-from quantaq_cli.utilities import safe_load
+from quantaq_cli.utilities import safe_load, determine_timestamp_column
 from quantaq_cli.variables import FLAG_DEFINITIONS, SUPPORTED_MODELS
+from quantaq_cli.console.commands.flag import (
+    echo_flag_table
+)
 
 
 def expunge_dataframe(df):
@@ -24,12 +25,14 @@ def expunge_dataframe(df):
 
     for label, value, cols in list_of_flags:
         mask = df["flag"] & value == value
-        n_affected = mask.sum()
-        pct_affected = round((n_affected / df.shape[0]) * 100.0, 2)
+        if not mask.any():
+            continue
 
         # NaN the necessary columns
-        if cols is None:
-            cols = df.columns
+        if cols == "all_columns":
+            tscol = determine_timestamp_column(df)
+            cols_to_keep = {tscol, "sn", "flag"}
+            cols = [c for c in df.columns if c not in cols_to_keep]
         elif len(cols) > 0:
             cols = [c for c in cols if c in df.columns]
     
@@ -37,60 +40,6 @@ def expunge_dataframe(df):
         df.loc[mask, cols] = np.nan
 
     return df
-
-def flag_summary(df):
-    """Create a table with a summary of flags.
-
-    Parameters
-    ----------
-    sensor_df
-        DataFrame with flags to summarize.
-
-    Returns
-    -------
-        A new DataFrame, suitable for human consumption.
-
-    TO DO: move this to flag.py after merging sc-20242
-    """
-    # force the flag column to be an int
-    df["flag"] = df["flag"].astype(int, errors='ignore')
-
-    rows = []
-    for name, value, cols in FLAG_DEFINITIONS:
-        mask = df["flag"] & value
-        num_naffected = mask.astype(bool).sum()
-        percent_affected = f"{100 * num_naffected / df.shape[0]:.1f}"
-        rows.append([name, int(value), num_naffected, percent_affected])
-
-    explained_df = pd.DataFrame(
-        rows,
-        columns=["FLAG", "FLAG VALUE", "# OCCURENCES", "% DATA"],
-    ).set_index("FLAG")
-    return explained_df
-
-def echo_flag_table(df):
-    """Print a table of flag statistics for a DataFrame.
-
-    Parameters
-    ----------
-    sensor_df
-        DataFrame to describe.
-
-    TO DO: move this to flag.py after merging sc-20242
-    """
-    explained_df = flag_summary(df)
-
-    table = Table()
-    for column in ["FLAG", *explained_df.columns]:
-        table.add_column(
-            column,
-            justify="right" if column != "FLAG" else "left",
-            style="bold",
-        )
-    for row in explained_df.itertuples():
-        table.add_row(*map(str, row))
-    rich.print(table)
-
 
 def expunge_command(file, output, **kwargs):
     verbose = kwargs.pop("verbose", False)
